Add profile page
Add profile page which displays all the following to the user:
i. The number of red-flag/intervention records that has been resolved
ii. The number of red-flag/intervention records that are yet to be resolved (in
draft or under investigation states).
iii. The number of red-flag/intervention records that has been rejected.
iv. List of all red-flag/intervention records.